feat: add metadata for these pages
