+++ src/app/[locale]/tools/han-nom-dictionaries/giup-doc-nom-va-han-viet/page.tsx
@@ -2,22 +2,8 @@ import { getTranslations } from "next-intl/server";
 import { Merriweather } from "next/font/google";
 import Entry from "./Entry";
 import DictionarySearchBar from "../DictionarySearchBar";
-import { Metadata } from "next";
 
 const merriweather = Merriweather({ weight: "300", subsets: ["vietnamese"] });
-
-export async function generateMetadata(): Promise<Metadata> {
-  const t = await getTranslations();
-
-  return {
-    title: `${t(
-      "Tools.han-nom-dictionaries.dictionaries.giup-doc-nom-va-han-viet.name"
-    )} | Digitizing Việt Nam`,
-    description: t(
-      "Tools.han-nom-dictionaries.dictionaries.giup-doc-nom-va-han-viet.description"
-    ),
-  };
-}
 
 export default async function DictionaryPage({
   searchParams,
