+++ src/app/[locale]/tools/han-nom-dictionaries/nguyen-trai-quoc-am-tu-dien/layout.tsx
@@ -3,7 +3,20 @@ import { getTranslations } from "next-intl/server";
 
 import BreadcrumbAndSearchBar from "@/components/layout/BreadcrumbAndSearchBar";
 import NavLink from "../NavLink";
+import { Metadata } from "next";
 
+export async function generateMetadata(): Promise<Metadata> {
+  const t = await getTranslations();
+
+  return {
+    title: `${t(
+      "Tools.han-nom-dictionaries.dictionaries.nguyen-trai-quoc-am-tu-dien.name"
+    )} | Digitizing Việt Nam`,
+    description: t(
+      "Tools.han-nom-dictionaries.dictionaries.nguyen-trai-quoc-am-tu-dien.description"
+    ),
+  };
+}
 export default async function DictionaryLayout({
   children,
   params: { locale },
